refactor logic and files
