+++ app/constants/domains.js
@@ -1,0 +1,10 @@
+const WOLOX_AR = 'wolox.ar';
+const WOLOX_CO = 'wolox.co';
+
+const validDomains = [WOLOX_AR, WOLOX_CO];
+
+module.exports = {
+  validDomains,
+  WOLOX_AR,
+  WOLOX_CO
+};
